app/src/main/java/com/nilhcem/droidconat/utils/Intents.java
package com.nilhcem.droidconat.utils;

import android.app.Activity;
import android.content.Context;
import android.content.Intent;
import android.net.Uri;
import android.support.annotation.NonNull;
import android.support.customtabs.CustomTabsIntent;
import android.support.v4.content.ContextCompat;

import com.nilhcem.droidconat.R;

public final class Intents {

    private Intents() {
        throw new UnsupportedOperationException();
    }

    public static boolean startUri(@NonNull Context context, @NonNull String url) {
        Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(url));
        if (intent.resolveActivity(context.getPackageManager()) != null) {
            context.startActivity(intent);
            return true;
        }
        return false;
    }

    public static void startExternalUrl(@NonNull Activity activity, @NonNull String url) {
        CustomTabsIntent intent = new CustomTabsIntent.Builder()
                .setShowTitle(true)
                .setToolbarColor(ContextCompat.getColor(activity, R.color.primary))
                .build();
        intent.launchUrl(activity, Uri.parse(url));
    }
}
